refactor: rename shortcuts
