+++ rest_framework_supertest/shortcuts/__init__.py
@@ -1,16 +1,16 @@
-from .automotive import *
-from .person import *
+from .automotives import *
+from .persons import *
 from .internet import *
 from .lorem import *
 from .isbn import *
-from .job import *
+from .jobs import *
 from .emojis import *
-from .credit_card import *
+from .credit_cards import *
 from .addresses import *
 from .phone_numbers import *
-from .currency import *
-from .geo import *
-from .file import *
+from .currencies import *
+from .geographicals import *
+from .files import *
 from .misc import *
 from .colors import *
 from .dates import *
