+++ rest_framework_supertest/shortcuts/automotives.py
