+++ rest_framework_supertest/shortcuts/currencies.py
